fanshawe-consultant-chatbot-web/requirements.txt
Flask==2.2.2
google-generativeai==0.1.0
numpy==1.21.6
pandas==1.3.5
requests==2.26.0